+++ src/compas_occ/conversions/__init__.py
@@ -13,11 +13,12 @@ Arrays
     :nosignatures:
 
     array1_from_floats1
-    array2_from_floats2
     array1_from_integers1
     array1_from_points1
+    array2_from_floats2
     array2_from_points2
     floats2_from_array2
+    harray1_from_points1
     points1_from_array1
     points2_from_array2
 
@@ -28,36 +29,28 @@ Primitives
     :toctree: generated/
     :nosignatures:
 
-    compas_point_to_occ_point
-    compas_vector_to_occ_vector
-    compas_vector_to_occ_direction
-    compas_plane_to_occ_plane
-    compas_plane_to_occ_ax2
-    compas_plane_to_occ_ax3
-    compas_line_to_occ_line
+    compas_circle_from_occ_circle
     compas_circle_to_occ_circle
-    compas_sphere_to_occ_sphere
-    compas_cylinder_to_occ_cylinder
     compas_cone_to_occ_cone
-    compas_torus_to_occ_torus
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
+    compas_cylinder_from_occ_cylinder
+    compas_cylinder_to_occ_cylinder
+    compas_frame_from_occ_ax3
+    compas_line_from_occ_line
+    compas_line_to_occ_line
+    compas_plane_from_occ_plane
     compas_point_from_occ_point
     compas_point_from_occ_point2d
+    compas_plane_to_occ_plane
+    compas_point_to_occ_point
+    compas_plane_to_occ_ax2
+    compas_plane_to_occ_ax3
+    compas_sphere_to_occ_sphere
+    compas_torus_to_occ_torus
+    compas_vector_from_occ_axis
     compas_vector_from_occ_vector
     compas_vector_from_occ_vector2d
-    compas_frame_from_occ_ax3
-
-Shapes
-======
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
+    compas_vector_to_occ_vector
+    compas_vector_to_occ_direction
 
 Transformations
 ===============
@@ -99,16 +92,21 @@ from .primitives import compas_point_from_occ_point  # noqa: F401
 from .primitives import compas_point_from_occ_point2d  # noqa: F401
 from .primitives import compas_vector_to_occ_vector  # noqa: F401
 from .primitives import compas_vector_to_occ_direction  # noqa: F401
+from .primitives import compas_vector_from_occ_axis  # noqa: F401
 from .primitives import compas_vector_from_occ_vector  # noqa: F401
 from .primitives import compas_vector_from_occ_vector2d  # noqa: F401
 from .primitives import compas_line_to_occ_line  # noqa: F401
+from .primitives import compas_line_from_occ_line  # noqa: F401
 from .primitives import compas_plane_to_occ_plane  # noqa: F401
+from .primitives import compas_plane_from_occ_plane  # noqa: F401
 from .primitives import compas_plane_to_occ_ax2  # noqa: F401
 from .primitives import compas_plane_to_occ_ax3  # noqa: F401
 from .primitives import compas_circle_to_occ_circle  # noqa: F401
+from .primitives import compas_circle_from_occ_circle  # noqa: F401
 from .primitives import compas_frame_from_occ_ax3  # noqa: F401
 from .primitives import compas_sphere_to_occ_sphere  # noqa: F401
 from .primitives import compas_cylinder_to_occ_cylinder  # noqa: F401
+from .primitives import compas_cylinder_from_occ_cylinder  # noqa: F401
 from .primitives import compas_cone_to_occ_cone  # noqa: F401
 from .primitives import compas_torus_to_occ_torus  # noqa: F401