+++ src/compas_occ/conversions/primitives.py
@@ -53,39 +53,41 @@ def compas_point_to_occ_point(point: Point) -> gp_Pnt:
     return gp_Pnt(*point)
 
 
-def compas_point_from_occ_point(cls: Type[Point], point: gp_Pnt) -> Point:
+def compas_point_from_occ_point(point: gp_Pnt, cls: Type[Point] = None) -> Point:
     """Construct a COMPAS point from an OCC point.
 
     Parameters
     ----------
-    cls : Type[:class:`~compas.geometry.Point`]
-        The type of COMPAS point.
     point : ``gp_Pnt``
         The OCC point.
+    cls : Type[:class:`~compas.geometry.Point`], optional
+        The type of COMPAS point.
 
     Returns
     -------
     :class:`~compas.geometry.Point`
 
     """
+    cls = cls or Point
     return cls(point.X(), point.Y(), point.Z())
 
 
-def compas_point_from_occ_point2d(cls: Type[Point], point: gp_Pnt2d) -> Point:
+def compas_point_from_occ_point2d(point: gp_Pnt2d, cls: Type[Point] = None) -> Point:
     """Construct a COMPAS point from an OCC 2D point.
 
     Parameters
     ----------
-    cls : Type[:class:`~compas.geometry.Point`]
-        The type of COMPAS point.
     point : ``gp_Pnt2d``
         The OCC point.
+    cls : Type[:class:`~compas.geometry.Point`], optional
+        The type of COMPAS point.
 
     Returns
     -------
     :class:`~compas.geometry.Point`
 
     """
+    cls = cls or Point
     return cls(point.X(), point.Y(), 0)
 
 
@@ -113,39 +115,43 @@ def compas_vector_to_occ_vector(vector: Vector) -> gp_Vec:
     return gp_Vec(*vector)
 
 
-def compas_vector_from_occ_vector(cls: Type[Vector], vector: gp_Vec) -> Vector:
+def compas_vector_from_occ_vector(vector: gp_Vec, cls: Type[Vector] = None) -> Vector:
     """Construct a COMPAS vector from an OCC vector.
 
     Parameters
     ----------
-    cls : Type[:class:`~compas.geometry.Vector`]
-        The type of COMPAS vector.
     vector : ``gp_Vec``
         The OCC vector.
+    cls : Type[:class:`~compas.geometry.Vector`], optional
+        The type of COMPAS vector.
 
     Returns
     -------
     :class:`~compas.geometry.Vector`
 
     """
+    cls = cls or Vector
     return cls(vector.X(), vector.Y(), vector.Z())
 
 
-def compas_vector_from_occ_vector2d(cls: Type[Vector], vector: gp_Vec2d) -> Vector:
+def compas_vector_from_occ_vector2d(
+    vector: gp_Vec2d, cls: Type[Vector] = None
+) -> Vector:
     """Construct a COMPAS vector from an OCC 2D vector.
 
     Parameters
     ----------
-    cls : Type[:class:`~compas.geometry.Vector`]
-        The type of COMPAS vector.
     vector : ``gp_Vec2d``
         The OCC vector.
+    cls : Type[:class:`~compas.geometry.Vector`], optional
+        The type of COMPAS vector.
 
     Returns
     -------
     :class:`~compas.geometry.Vector`
 
     """
+    cls = cls or Vector
     return cls(vector.X(), vector.Y(), 0)
 
 
@@ -173,22 +179,43 @@ def compas_vector_to_occ_direction(vector: Vector) -> gp_Dir:
     return gp_Dir(*vector)
 
 
-def compas_vector_from_occ_direction(cls: Type[Vector], vector: gp_Dir) -> Vector:
+def compas_vector_from_occ_direction(
+    vector: gp_Dir, cls: Type[Vector] = None
+) -> Vector:
     """Construct a COMPAS vector from an OCC direction.
 
     Parameters
     ----------
-    cls : Type[:class:`~compas.geometry.Vector`]
-        The type of COMPAS vector.
     vector : ``gp_Dir``
         The OCC direction.
+    cls : Type[:class:`~compas.geometry.Vector`], optional
+        The type of COMPAS vector.
 
     Returns
     -------
     :class:`~compas.geometry.Vector`
 
     """
+    cls = cls or Vector
     return cls(vector.X(), vector.Y(), vector.Z())
+
+
+def compas_vector_from_occ_axis(axis: gp_Ax1, cls: Type[Vector] = None) -> Vector:
+    """Convert an OCC axis to a COMPAS vector.
+
+    Parameters
+    ----------
+    axis : ``gp_Ax1``
+        The OCC axis.
+    cls : Type[:class:`compas.geometry.Vector`], optional
+        The type of COMPAS vector.
+
+    Returns
+    -------
+    :class:`~compas.geometry.Vector`
+
+    """
+    return compas_vector_from_occ_direction(axis.Direction(), cls=cls)
 
 
 def compas_axis_to_occ_axis(axis: Tuple[Point, Vector]) -> gp_Ax1:
@@ -263,6 +290,14 @@ def compas_line_to_occ_line(line: Line) -> gp_Lin:
     )
 
 
+def compas_line_from_occ_line(lin: gp_Lin, cls: Type[Line] = None) -> Line:
+    """Convert an OCC line to a COMPAS line."""
+    cls = cls or Line
+    point = compas_point_from_occ_point(lin.Location())
+    vector = compas_vector_from_occ_direction(lin.Direction())
+    return cls(point, point + vector)
+
+
 def compas_plane_to_occ_plane(plane: Plane) -> gp_Pln:
     """Convert a COMPAS plane to an OCC plane.
 
@@ -279,6 +314,15 @@ def compas_plane_to_occ_plane(plane: Plane) -> gp_Pln:
     return gp_Pln(
         compas_point_to_occ_point(plane.point),
         compas_vector_to_occ_direction(plane.normal),
+    )
+
+
+def compas_plane_from_occ_plane(pln: gp_Pln, cls: Type[Plane] = None) -> Plane:
+    """Convert an OCC plane to a COMPAS plane."""
+    cls = cls or Plane
+    return cls(
+        compas_point_from_occ_point(pln.Location()),
+        compas_vector_from_occ_axis(pln.Axis()),
     )
 
 
@@ -320,25 +364,26 @@ def compas_plane_to_occ_ax3(plane: Plane) -> gp_Ax3:
     )
 
 
-def compas_frame_from_occ_ax3(cls: Type[Frame], position: gp_Ax3) -> Frame:
+def compas_frame_from_occ_ax3(position: gp_Ax3, cls: Type[Frame] = None) -> Frame:
     """Construct a COMPAS frame from an OCC position.
 
     Parameters
     ----------
-    cls : Type[:class:`~compas.geometry.Frame`]
-        The type of COMPAS frame.
     position : ``gp_Ax3``
         The OCC position.
+    cls : Type[:class:`~compas.geometry.Frame`], optional
+        The type of COMPAS frame.
 
     Returns
     -------
     :class:`~compas.geometry.Frame`
 
     """
+    cls = cls or Frame
     return cls(
-        compas_point_from_occ_point(Point, position.Location()),
-        compas_vector_from_occ_direction(Vector, position.XDirection()),
-        compas_vector_from_occ_direction(Vector, position.YDirection()),
+        compas_point_from_occ_point(position.Location()),
+        compas_vector_from_occ_direction(position.XDirection()),
+        compas_vector_from_occ_direction(position.YDirection()),
     )
 
 
@@ -356,6 +401,13 @@ def compas_circle_to_occ_circle(circle: Circle) -> gp_Circ:
 
     """
     return gp_Circ(compas_plane_to_occ_ax2(circle.plane), circle.radius)
+
+
+def compas_circle_from_occ_circle(circ: gp_Circ, cls: Type[Circle] = None) -> Circle:
+    """Construct a COMPAS circle from an OCC circle."""
+    cls = cls or Circle
+    point, vector = compas_axis_from_occ_axis(circ.Axis())
+    return cls(Plane(point, vector), circ.Radius())
 
 
 def compas_sphere_to_occ_sphere(sphere: Sphere) -> gp_Sphere:
@@ -389,6 +441,17 @@ def compas_cylinder_to_occ_cylinder(cylinder: Cylinder) -> gp_Cylinder:
 
     """
     return gp_Cylinder(compas_plane_to_occ_ax3(cylinder.plane), cylinder.circle.radius)
+
+
+def compas_cylinder_from_occ_cylinder(
+    cylinder: gp_Cylinder, cls: Type[Cylinder] = None
+) -> Cylinder:
+    """Convert an OCC cylinder to a COMPAS cylinder."""
+    circle = Circle(
+        Plane(*compas_axis_from_occ_axis(cylinder.Axis())), cylinder.Radius()
+    )
+    cls = cls or Cylinder
+    return cls(circle, 1.0)
 
 
 def compas_cone_to_occ_cone(cone: Cone) -> gp_Cone: